requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
tzdata==2024.2
djangorestframework~=3.15.2
django-mptt~=0.16.0

openpyxl~=3.1.5